fastapi_app/requirements.txt
aerich==0.8.1
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.8.0
asyncclick==8.1.8.0
asyncpg==0.30.0
bcrypt==4.2.1
click==8.1.8
dictdiffer==0.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.8
h11==0.14.0
idna==3.10
iso8601==2.1.0
passlib==1.7.4
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pypika-tortoise==0.5.0
python-dotenv==1.0.1
pytz==2025.1
sniffio==1.3.1
starlette==0.45.3
tomlkit==0.13.2
tortoise-orm==0.24.0
typing_extensions==4.12.2
uvicorn==0.34.0
websockets==15.0
